+++ examples/misc/raptor.cfg
@@ -1,0 +1,33 @@
+
+{
+    "workload"       : {
+        "total"      : "1024 * 64",
+        "timeout"    : 1
+    },
+
+    "runtime"        : 120,
+    "cpn"            : 8,
+    "gpn"            : 2,
+
+    "n_masters"      :   2,
+    "n_workers"      :   16,
+
+    "master"         : "raptor_master.py",
+    "worker"         : "raptor_worker.py",
+
+    "msg_batch"      : 10,
+
+
+    "pilot_descr"    : {
+        "resource"   : "local.localhost"
+    },
+
+    "master_descr"   : {
+        "executable" : "python3"
+    },
+
+    "worker_descr"   : {
+        "executable" : "python3"
+    }
+}
+
