+++ examples/misc/raptor.frontera.cfg
@@ -8,8 +8,8 @@
 
     "total"     : "1024",
 
-    "master"    : "task_overlay_master.py",
-    "worker"    : "task_overlay_worker.py",
+    "master"    : "raptor_master.py",
+    "worker"    : "raptor_worker.py",
 
     "pilot_descr"  : {
         "resource"   : "xsede.frontera",
