+++ examples/misc/raptor.py
