+++ examples/misc/raptor.summit.cfg
@@ -14,8 +14,8 @@
     "nodes"          :   8,
     "n_masters"      :   1,
 
-    "master"         : "task_overlay_master.py",
-    "worker"         : "task_overlay_worker.py",
+    "master"         : "raptor_master.py",
+    "worker"         : "raptor_worker.py",
 
     "msg_batch"      : 10,
