--- examples/misc/task_overlay.cfg
@@ -1,39 +1,0 @@
-
-{
-    "workload"       : {
-        "total"      : "1024 * 64",
-        "timeout"    : 1
-    },
-
-    "runtime"        : 120,
-    "cpn"            : 8,
-    "gpn"            : 2,
-
-    "n_masters"      :   2,
-    "n_workers"      :   16,
-
-    "master"         : "task_overlay_master.py",
-    "worker"         : "task_overlay_worker.py",
-
-    "msg_batch"      : 10,
-
-
-    "pilot_descr"    : {
-        "resource"   : "local.localhost"
-    },
-
-    "master_descr"   : {
-        "executable" : "python3",
-        "pre_exec"   : [". $RP_PILOT_SANDBOX/deactivate",
-                        ". $HOME/.miniconda3/etc/profile.d/conda.sh",
-                        "conda activate tmp"]
-    },
-
-    "worker_descr"   : {
-        "executable" : "python3",
-        "pre_exec"   : [". $RP_PILOT_SANDBOX/deactivate",
-                        ". $HOME/.miniconda3/etc/profile.d/conda.sh",
-                        "conda activate tmp"]
-    }
-}
-
