+++ src/radical/pilot/__init__.py
@@ -37,10 +37,10 @@ from . import umgr
 from . import pmgr
 from . import agent
 
-from .agent import Agent_0
-from .agent import Agent_n
+from .agent  import Agent_0
+from .agent  import Agent_n
 
-from .task_overlay import Master, Worker
+from .raptor import Master, Worker
 
 
 # ------------------------------------------------------------------------------
